backend/requirements.txt
# ============================================
# Multi-Agent Chatbot - Requirements
# ============================================

# -------------------- Core FastAPI & Server --------------------
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
pydantic==2.11.9
pydantic_core==2.33.2
python-multipart==0.0.20

# -------------------- Database - MongoDB --------------------
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# -------------------- Google OAuth & API --------------------
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.185.0
PyJWT==2.10.1
python-jose[cryptography]==3.5.0
cryptography==46.0.1
ecdsa==0.19.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
cachetools==6.2.1
oauthlib==3.3.1
requests-oauthlib==2.0.0
uritemplate==4.2.0
googleapis-common-protos==1.70.0
protobuf==6.33.0
proto-plus==1.26.1
httplib2==0.31.0
pyparsing==3.2.5

# -------------------- OpenAI API --------------------
openai==2.2.0
httpx==0.28.1
httpcore==1.0.9
distro==1.9.0
jiter==0.11.0

# -------------------- LangChain & LangGraph (Multi-Agent Orchestration) --------------------
langchain==0.3.27
langchain-core==0.3.78
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-checkpoint==2.1.2
langsmith==0.4.32
jsonpatch==1.33
jsonpointer==3.0.0
tenacity==9.1.2

# -------------------- Environment & Configuration --------------------
python-dotenv==1.1.1
PyYAML==6.0.3

# -------------------- HTTP & Network --------------------
requests==2.32.5
aiohttp==3.12.15
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
frozenlist==1.8.0
multidict==6.6.4
propcache==0.4.0
yarl==1.21.0
attrs==25.3.0

# -------------------- Utilities --------------------
typing_extensions==4.15.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
h11==0.16.0
idna==3.10
sniffio==1.3.1
urllib3==2.5.0
packaging==25.0

# -------------------- Development & Testing (Optional) --------------------
# Uncomment these for development
# pytest==8.4.2
# black==25.9.0
# flake8==7.3.0
# isort==6.0.1
# mypy==1.18.2

# -------------------- JSON Schema & Validation --------------------
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.36.2
rpds-py==0.27.1

# -------------------- Additional Dependencies --------------------
# Auto-installed by dependencies above but listed for clarity
pycparser==2.23
cffi==2.0.0
six==1.17.0
